videoplayer created and configured
