+++ Navigation/Music/View/MusicViewController.swift
@@ -57,13 +57,13 @@ class MusicViewController: UIViewController {
         
         playerView.snp.makeConstraints { make in
             make.leading.trailing.equalToSuperview()
-            make.bottom.equalToSuperview().inset(50)
-            make.height.equalTo(50)
+            make.bottom.equalToSuperview().inset(80)
+            make.height.equalTo(80)
         }
     }
 }
 
-extension MusicViewController: UITableViewDelegate, UITableViewDataSource {
+extension MusicViewController: UITableViewDataSource {
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: String(describing: MusicTableViewCell.self), for: indexPath) as? MusicTableViewCell,
@@ -73,7 +73,7 @@ extension MusicViewController: UITableViewDelegate, UITableViewDataSource {
     }
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return viewModel?.tracks.count ?? 0
+        return MusicViewModel.tracks.count
     }
 
     func numberOfSections(in tableView: UITableView) -> Int {
@@ -81,72 +81,15 @@ extension MusicViewController: UITableViewDelegate, UITableViewDataSource {
     }
     
     func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        return "Список треков"
+        return "   Список треков"
     }
     
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        
-        playerView.playSelectedTrack(forIndex: indexPath.row)
-        
-        tableView.deselectRow(at: indexPath, animated: true)
-    }
 }
-
-
-
-
-
-class MusicTableViewCell: UITableViewCell {
     
-    // MARK: PROPERTIES ============================================================================
+extension MusicViewController: UITableViewDelegate {
 
-    private lazy var trackNameLabel: UILabel = {
-        let label = UILabel()
-        return label
-    }()
-    
-    private lazy var musicIcon: UIImageView = {
-        let image = UIImageView()
-        image.image = UIImage(systemName: "music.note", withConfiguration: UIImage.SymbolConfiguration(pointSize: 32))?.withTintColor(.black, renderingMode: .alwaysOriginal)
-        return image
-    }()
-    
-    // MARK: INITIALIZATORS ============================================================================
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        contentView.addSubviews(trackNameLabel, musicIcon)
-        setupLayout()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: METHODS ===================================================================================
-
-    func setupLayout() {
-                
-        musicIcon.snp.makeConstraints { make in
-            make.leading.top.bottom.equalToSuperview().inset(12)
-        }
-        
-        trackNameLabel.snp.makeConstraints { make in
-            make.top.bottom.equalToSuperview().inset(12)
-            make.leading.equalTo(musicIcon.snp.trailing).offset(12)
-        }
-    }
-    
-    func setConfigureOfCell(model: MusicViewModel, index: Int) {
-        
-        var currentTrackName: String {
-            get {
-                let singer = Array(model.tracks.values)[index]
-                let track = Array(model.tracks.keys)[index]
-                return "\(singer) - \(track)"
-            }
-        }
-        
-        trackNameLabel.text = currentTrackName
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        playerView.playSelectedTrack(forIndex: indexPath.row)
+        tableView.deselectRow(at: indexPath, animated: true)
     }
 }
