+++ Navigation/Video/Coordinator/VideoCoordinator.swift
@@ -12,7 +12,16 @@ final class VideoCoordinator {
         let viewModel = VideoViewModel()
         let viewController = VideoViewController(model: viewModel, coordinator: coordinator)
         viewController.view.backgroundColor = .secondarySystemGroupedBackground
-        viewController.title = "Music"
+        viewController.title = "Video"
         return viewController
+    }
+}
+
+final class VideoPlayerCoordinator {
+    func showDetail(navCon: UINavigationController?, coordinator: VideoPlayerCoordinator, key: String) {
+        let viewController = VidePlayerViewController(videoKey: key)
+        viewController.view.backgroundColor = .systemGray5
+        viewController.title = VideoViewModel.videos[key]
+        navCon?.pushViewController(viewController, animated: true)
     }
 }
