+++ Navigation/Video/ViewModel/VideoViewModel.swift
@@ -7,16 +7,19 @@
 
 import UIKit
 import AVKit
-import youtube_ios_player_helper
 
 
 final class VideoViewModel {
-        
     
+    static var videos: [String: String] = [
+        "EbHGS_bVkXY" : "Rammstein - Zeit",
+        "hBTNyJ33LWI" : "Rammstein - Zick Zack",
+        "thJgU9jkdU4" : "Rammstein - Dicke Titten",
+        "ONj9cvHCado" : "Rammstein - Angst",
+    ]
+        
 //    private lazy var streamURL = URL(string: "https://devstreaming-cdn.apple.com/videos/streaming/examples/bipbop_adv_example_hevc/master.m3u8")!
     private lazy var streamURL = URL(string: "https://www.youtube.com/watch?v=KfLxG2AdfXw")!
-
-
 
     private lazy var localURL: URL = {
         let path = Bundle.main.path(forResource: "test", ofType: "mp4")!
